+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.distsql.query.DistSQLResultSet
@@ -15,5 +15,5 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.proxy.backend.text.distsql.rql.DataSourceQueryResultSet
-org.apache.shardingsphere.proxy.backend.text.distsql.rql.SingleTableRuleQueryResultSet
+org.apache.shardingsphere.proxy.backend.text.distsql.rql.resource.DataSourceQueryResultSet
+org.apache.shardingsphere.proxy.backend.text.distsql.rql.rule.SingleTableQueryResultSet
